Fix version to 1.0.0-SNAPSHOT
ctrueden observed that incrementing the tutorial GAV is unnecessary.
Nothing will ever rely on the tutorials, so there is no need to ever
"release" new versions of them.

So let's fix the version. Labeled as a SNAPSHOT because otherwise we
would be bound to constantly increasing the version with each commit.